+++ client/src/pages/spectrogram/freq-selector.tsx
