+++ client/src/pages/spectrogram/ruler-side.tsx
