+++ client/src/pages/spectrogram/ruler-top.tsx
