+++ client/src/pages/spectrogram/scroll-bar.tsx
