+++ client/src/pages/spectrogram/settings-pane.tsx
