+++ client/src/pages/spectrogram/time-selector-minimap.tsx
